requirement.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
load-dotenv==0.1.0
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
django_filters=24.3
djangorestframework-simplejwt=5.3.1
django_extensions=3.2.3
coverage=7.6.7